# Conflicts:
#	car-rental/src/main/java/com/nextcar/carrental/controller/CarController.java
#	car-rental/src/main/java/com/nextcar/carrental/service/CarService.java
#	car-rental/src/main/resources/static/about.html
#	car-rental/src/main/resources/static/profile.html

+++ car-rental/src/main/java/com/nextcar/carrental/entity/Car.java
@@ -10,7 +10,7 @@ public class Car {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @NotBlank(message = "Märke är obligatoriskt")
     @Size(min = 2, max = 50, message = "Märke måste vara mellan 2-50 tecken")
@@ -75,8 +75,8 @@ public class Car {
     private String status = "ACTIVE";
 
     // Getters and setters
-    public Integer getId() { return id; }
-    public void setId(Integer id) { this.id = id; }
+    public Long getId() { return id; }
+    public void setId(Long id) { this.id = id; }
 
     public String getBrand() { return brand; }
     public void setBrand(String brand) { this.brand = brand; }
@@ -106,7 +106,7 @@ public class Car {
     public void setPrice(BigDecimal price) { this.price = price; }
 
     // Hjälpmetod för att hämta categoryId direkt
-    public Integer getCategoryId() {
+    public Long getCategoryId() {
         return category != null ? category.getId() : null;
     }
